requirements.txt
djangorestframework
django-cors-headers
drf-yasg
pytest
pytest-django
pandas
